requirements.txt
flask==2.0.1
gunicorn==20.1.0
werkzeug==2.0.3
psycopg2-binary==2.9.6